Разбил urls и views по папкам
