+++ source/api/views/api.py
@@ -365,136 +365,3 @@ class LikeCreateView(generics.CreateAPIView):
 class LikeDestroyView(generics.DestroyAPIView):
     queryset = Like.objects.all()
     serializer_class = LikeDeleteSerializer
-
-
-
-# ---------------- Регистрация ----------------
-"""class IsProfileOwner(permissions.BasePermission):
-    def has_object_permission(self, request, view, obj):
-        return request.user == obj.user"""
-
-class RegisterView(generics.CreateAPIView):
-    serializer_class = ProfileCreateSerializer
-    
-    @transaction.atomic
-    def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        
-        try:
-            profile = serializer.save()
-            
-            return Response(
-                {
-                    "status": "success",
-                    "user_id": profile.user.id,
-                    "profile_id": profile.id,
-                    "message": "Регистрация прошла успешно"
-                },
-                status=status.HTTP_201_CREATED
-            )
-
-        except Exception as e:
-            if "unique constraint" in str(e).lower():
-                return Response(
-                    {"error": "Пользователь с такими данными уже существует"},
-                    # TO DO Добавить файл с ошибками
-                    status=status.HTTP_400_BAD_REQUEST
-                )
-            return Response(
-                {"error": str(e)},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-
-
-# ---------------- Смена пароля ----------------
-# TO DO Возможно стоит добавить сериализатор для изменения пароля
-# Если добавлять, то явно в new_password задать min_length=8 и max_length=128
-class ChangePasswordView(APIView):
-    permission_classes = [IsAuthenticated]  
-
-    def post(self, request):
-        user = request.user
-        old_password = request.data.get('old_password')
-        new_password = request.data.get('new_password')
-
-
-        if not user.check_password(old_password):
-            return JsonResponse({'detail': 'Неверный старый пароль.'}, status=400)
-
-        if len(new_password) < 8:
-            return JsonResponse({'detail': 'Пароль должен быть не менее 8 символов.'}, status=400)
-
-        user.set_password(new_password)
-        user.save()
-        update_session_auth_hash(request, user)
-        return JsonResponse({'success': True})
-
-
-# ---------------- Логин ----------------
-# TO DO добавить сериализатор для логина
-# Поменять названия
-def login_view(request):
-    if request.method == 'POST':
-        form = AuthenticationForm(request, data=request.POST)
-        if form.is_valid():
-            user = form.get_user()
-            login(request, user)
-            messages.success(request, f"Добро пожаловать, {user.username}!")
-            return redirect('index')
-        else:
-            messages.error(request, "Неверное имя пользователя или пароль.")
-    else:
-        form = AuthenticationForm()
-
-    return render(request, 'meetly/login.html', {'form': form})
-
-
-# ---------------- Страницы ----------------
-# TO DO Добавить файл бекенда для аутентификации
-def index(request):
-    return render(request, 'meetly/index.html')
-
-
-def change_password_page(request):
-    return render(request, 'meetly/change-password.html')
-
-
-def friends_page(request):
-    profile = request.user.profile 
-    friends_profiles = profile.friends.all() 
-    friends = User.objects.filter(profile__in=friends_profiles)  
-    others = User.objects.exclude(profile__in=friends_profiles).exclude(id=request.user.id)  
-    context = {
-        'friends': friends,
-        'others': others,
-    }
-    return render(request, 'friends.html', context)
-
-
-class AddFriendView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request):
-        user_id = request.data.get('profile_id')
-        if not user_id:
-            return Response({'error': 'ID профиля обязателен'}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            user = User.objects.get(id=user_id)
-            friend_profile = user.profile
-            profile = request.user.profile
-            if friend_profile in profile.friends.all():
-                return Response({'error': 'Этот пользователь уже в списке ваших друзей.'}, status=status.HTTP_400_BAD_REQUEST)
-            profile.friends.add(friend_profile)
-
-            return Response({'success': True}, status=status.HTTP_200_OK)
-
-        except Profile.DoesNotExist:
-            return Response({'error': 'Профиль не найден'}, status=status.HTTP_404_NOT_FOUND)
-
-        except Exception as e:
-            import logging
-            logger = logging.getLogger(__name__)
-            logger.exception("Ошибка при добавлении в друзья")
-            return Response({'error': 'Произошла ошибка на сервере'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
